Numa.jl

Numa.jl

Documentation for the Numa library

Numa.derivative — Method.

Function to be eliminated in the future. Compute the numder-th derivative of a monomial at a set of 1D point

source

Numa.evaluate! — Method.

First axis of v for dofs, second for points

source

Numa.gradient — Method.

Returns a MultivariatePolynomialBasis{TG,D} where TG is a type whose rank is one unit grater than the one of T

source

Numa.gradshfsps — Method.

Compute shape functions gradients (to be re-implemented)

source

Numa.shfsps — Method.

Compute shape functions in a set of points

source

Numa.FESpace — Type.

FE Space structure, where only one RefFE is possible in the whole mesh (to be improved in the future)

source

Numa.LagrangianRefFE — Type.

Reference Finite Element a la Ciarlet, i.e., it relies on a local function (polynomial) space, an array of nodes (DOFs), and a polytope (cell topology). The rank of the approximating field can be arbitrary. The current implementation relies on the prebasis (e.g., monomial basis of polynomials) and a change-of-basis (using the node array) to generate the canonical basis, i.e., the shape functions.

source

Numa.MPoint — Type.

The mutable version of Point{D}

source

Numa.MTensorValue — Type.

Mutable version of TensorValue{D,DD}

source

Numa.MVectorValue — Type.

Mutable version of VectorValue{D}

source

Numa.MultivariatePolynomialBasis — Type.

Abstract type representing a multivariate polynomial basis with value of type T in a coordinate space of D dimensions

source

Numa.NFace — Type.

n-face of the polytope, i.e., any polytope of lower dimension (n) representing its boundary and the polytope itself (for n equal to the space dimension)

source

Numa.NodesArray — Type.

Array of nodes for a given polytope and order

source

Numa.Point — Type.

Type representing a point of D dimensions

source

Numa.Polytope — Type.

Aggregation of all n-faces that compose the polytope boundary and the polytope itself, the classification of n-faces with respect to their dimension and type

source

Numa.Quadrature — Type.

Abstract type representing a quadrature rule on a Polytope in a space of D dimensions

source

Numa.ScalarValue — Type.

Type representing a scalar value

source

Numa.TensorProductPolynomialBasis — Type.

Multivariate polynomial basis obtained as tensor product of univariate polynomial basis per dimension

source

Numa.TensorProductPolynomialBasis — Method.

Provide a TensorProductPolynomialBasis for a vector order providing the order per dimension

source

Numa.TensorProductQuadrature — Type.

Tensor product quadrature rule (nodes and weights) on a hyper cube [-1,1]^D

source

Numa.TensorProductQuadratureOld — Type.

Tensor product quadrature rule (nodes and weights) integrating exactly 2´order´-1 polynomials

source

Numa.TensorValue — Type.

Type representing a tensor value of dimension D

source

Numa.UnivariateMonomialBasis — Type.

Univariate monomial basis of a given order

source

Numa.UnivariateMonomialBasis — Method.

Evaluate univariate monomial basis in a set of 1D points

source

Numa.UnivariatePolynomialBasis — Type.

Abstract basis of univariate polynomials

source

Numa.UnivariatePolynomialBasis — Method.

Create 1-dim univariate polynomial basis of UnivariateMonomialBasis type

source

Numa.VectorValue — Type.

Type representing a vector value of dimension D

source

Numa.RefFE — Type.

Abstract Reference Finite Element

source