Gridap.CellData

Gridap.CellData — Module

The exported names are

CellDatum

CellDof

CellField

CellPoint

CellQuadrature

CellState

DiracDelta

DomainContribution

DomainStyle

GenericCellField

Integrand

Measure

PhysicalDomain

ReferenceDomain

add_contribution!

attach_constraints_cols

attach_constraints_rows

attach_dirichlet

change_domain

compute_cell_points_from_vector_of_points

cross

det

dot

double_contraction

get_cell_measure

get_cell_points

get_contribution

get_data

get_domains

get_normal_vector

get_physical_coordinate

get_triangulation

gradient

identity_constraints

inner

jump

make_inverse_table

mean

num_domains

outer

symmetric_part

test_cell_datum

tr

update_state!

×

∇

∇∇

∫

⊗

⊙

⋅

⋅²

⋅¹

Gridap.CellData.CellDatum — Type

Data associated with the cells of a Triangulation. CellDatum objects behave as if they are defined in the physical space of the triangulation. But in some cases they are implemented as reference quantities plus some transformation to the physical domain.

Gridap.CellData.CellDof — Type

Gridap.CellData.CellField — Type

Gridap.CellData.CellPoint — Type

A single point or an array of points on the cells of a Triangulation CellField objects can be evaluated efficiently at CellPoint instances.

Gridap.CellData.CellQuadrature — Type

Gridap.CellData.CellState — Type

This can be used as a CellField as long as one evaluates it on the stored CellPoint.

Gridap.CellData.DomainContribution — Type

Gridap.CellData.DomainStyle — Type

Trait that signals if a CellDatum type is implemented in the physical or the reference domain

Gridap.CellData.DomainStyle — Method

Tell if the stored array is in the reference or physical domain

Gridap.CellData.GenericCellField — Type

Gridap.Arrays.get_array — Method

Get the raw array of cell data defined in the physical space.

Gridap.CellData.change_domain — Method

Gridap.CellData.change_domain — Method

Change the underlying data to the target domain

Gridap.CellData.distance — Method

dist = distance(polytope::ExtrusionPolytope, inv_cmap::Field, x::Point)

Calculate distance from point x to the polytope. The polytope is given by its type and by the inverse cell map, i.e. by the map from the physical to the reference space.

Positive distances are outside the polytope, negative distances are inside the polytope.

The distance is measured in an unspecified norm, currently the L∞ norm.

Gridap.CellData.get_cell_measure — Method

Contributions added to the cells of the background Triangulation.

Gridap.CellData.get_cell_points — Method

Gridap.CellData.get_data — Method

Get the stored array of cell-wise data. It can be defined in the physical or the reference domain.

Gridap.CellData.get_physical_coordinate — Method

get_physical_coordinate(trian::Triangulation)

In contrast to getcellmap, the returned object:

is a CellField

its gradient is the identity tensor

Gridap.CellData.test_cell_datum — Method

Gridap.Geometry.get_triangulation — Method

Return the underlying Triangulation object