Gridap.Io

Gridap.Io — Module

The exported names in this module are:

from_bson_file

from_dict

from_jld2_file

from_json

from_json_file

to_bson_file

to_dict

to_jld2_file

to_json

to_json_file

Gridap.Io.check_dict — Method

check_dict(::Type{T},dict::Dict) where T

Check validity of a dictionary dict for an object of type T. It runs successfully if the dictionary is valid for a particular type or throws an error in any other case.

Gridap.Io.from_bson_file — Method

function from_bson_file(::Type{T},s::AbstractString) where T

Gridap.Io.from_dict — Method

from_dict(::Type{T},dict::Dict) where T

De-serialize an object of type T from the dictionary dict. Values stored into this dictionary must be of any native Julia data type (Real, Integer, String, etc.) Dictionary keys are Symbols.

Gridap.Io.from_jld2_file — Function

function from_jld2_file(filename::AbstractString,dataset::AbstractString="data")

Loads an object from a JLD2 file given its filename and, optionally, the dataset name. The dataset specifies the name and the root location of the data inside the generated JLD2 file.

Gridap.Io.from_jld2_file — Method

function from_jld2_file(::Type{T},filename::AbstractString,dataset::AbstractString="data") where T

Loads an object from a JLD2 file given its filename and, optionally, the dataset name. The dataset specifies the name and the root location of the data inside the generated JLD2 file. Checks if the returned object is of the expected Type{T}, if not return error.

Gridap.Io.from_json — Method

from_json(::Type{T},s::AbstractString) where T

Gridap.Io.from_json_file — Method

from_json_file(::Type{T},s::AbstractString) where T

Gridap.Io.to_bson_file — Method

to_bson_file(object,filename)

Gridap.Io.to_dict — Method

to_dict(object) -> Dict

Serialize object into a dictionary of type Dict{Symbol,Any}. Values stored into this dictionary must be of any native Julia data type (Real, Integer, String, etc.) Dictionary keys are Symbols.

Gridap.Io.to_jld2_file — Function

function to_jld2_file(object,filename::AbstractString,dataset::AbstractString="data")

Stores an object to a JLD2 file given its filename and, optionally, the dataset name. The dataset specifies the name and the root location of the data inside the JLD2 file.

Gridap.Io.to_json — Method

to_json(object)

Gridap.Io.to_json_file — Method

to_json_file(object,filename)