TikzPictures.jl extension

Gridap provides a simple way of exporting geometries into LaTeX through Tikz, by interfacing with TikzPictures.jl:

TikzPictures.TikzPicture — Method

TikzPictures.TikzPicture(model::DiscreteModel;kwargs...)
TikzPictures.TikzPicture(grid::Grid;kwargs...)
TikzPictures.TikzPicture(topo::GridTopology;kwargs...)
TikzPictures.TikzPicture(topo::Polytope;kwargs...)

Returns a TikzPicture for the given model, grid, or topology.

Available keyword arguments include:

draw_nodes: whether to draw visible nodes (default: true)

draw_labels: whether to add labels (default: true)

edge_style: TikZ style for edges (default: "[ ]")

node_style: TikZ style for nodes (default: "[circle, fill=black, inner sep=0pt, minimum size=0.1cm]")

label_style: TikZ style for labels (default: "[ ]")

label_offset: offset for labels, w.r.t. nodes (default: 0.0)

label_dim: dimension of the labeled faces (default: 0, i.e nodes)

as well as any TikzPicture-specific keyword arguments.

This is how our table of polytopes was generated.