Numa.jl

Numa.jl

Documentation for the Numa library

Numa.evaluate! — Method.

First axis of v for dofs, second for points

source

Numa.gradient — Method.

Returns another CellBasis object that represents the gradient TG is a value whose rank is one order grater than the one of T

source

Numa.gradient — Method.

Returns another CellField object that represents the gradient. TG is a value whose rank is one order grater than the one of T

source

Numa.gradient — Method.

Returns a MultivariatePolynomialBasis{TG,D} where TG is a type whose rank is one unit grater than the one of T

source

Numa.gradshfsps — Method.

Compute shape functions gradients (to be re-implemented)

source

Numa.mapderivatives — Method.

Returns another CellBasis, whose spatial derivatives are respect to the coordinates of the range space of geomap

source

Numa.shfsps — Method.

Compute shape functions in a set of points

source

Numa.CellArray — Type.

Abstract type representing an iterable collection of Arrays{T,N}, where each array is associated with a cell.

source

Numa.CellBasis — Type.

T is for the type of value D is for the dimension of the domain

source

Numa.CellBasisValues — Type.

Abstract type that represents a function basis with value of type T evaluated at a collection of points in each cell

source

Numa.CellFieldValues — Type.

Abstract type that represents a field with value of type T evaluated at a collection of points in each cell

source

Numa.CellMatrices — Type.

Abstract type that represents a matrix of value T associated with a collection of points in each cell (typically the cell matrix at the quadrature points)

source

Numa.CellPoints — Type.

An array of points for each cell

source

Numa.CellQuadrature — Type.

Abstract type representing a collection of quadratures, one for each cell

source

Numa.CellScalars — Type.

Abstract type that represents a scalar of value T associated with a collection of points in each cell

source

Numa.CellValues — Type.

An array of values for each cell

source

Numa.CellVectors — Type.

Abstract type that represents a vector of value T associated with a collection of points in each cell (typically the cell rhs vector at the quadrature points)

source

Numa.ConstantCellArray — Type.

Concrete implementation of CellArray, where the same array is associated to all cells. Typically, this is useful for discretizations with a single cell type.

source

Numa.ConstantCellQuadrature — Type.

A concrete implementation of CellQuadrature for the particular case that all cells have the same quadrature

source

Numa.FESpace — Type.

FE Space structure, where only one RefFE is possible in the whole mesh (to be improved in the future)

source

Numa.IndexableCellArray — Type.

Abstract type representing an indexable CellArray. By implementing a concrete IndexableCellArray, one automatically gets an type that is also iterable

source

Numa.IntegrationDomain — Type.

This is the very minimum needed to describe the domain for numerical integration

source

Numa.IntegrationMesh — Type.

Minimal interface for a mesh used for numerical integration

source

Numa.LagrangianRefFE — Type.

Reference Finite Element a la Ciarlet, i.e., it relies on a local function (polynomial) space, an array of nodes (DOFs), and a polytope (cell topology). The rank of the approximating field can be arbitrary. The current implementation relies on the prebasis (e.g., monomial basis of polynomials) and a change-of-basis (using the node array) to generate the canonical basis, i.e., the shape functions.

source

Numa.MPoint — Type.

The mutable version of Point{D}

source

Numa.MVectorValue — Type.

Mutable version of VectorValue

source

Numa.MultivariatePolynomialBasis — Type.

Abstract type representing a multivariate polynomial basis with value of type T in a coordinate space of D dimensions

source

Numa.NFace — Type.

n-face of the polytope, i.e., any polytope of lower dimension (n) representing its boundary and the polytope itself (for n equal to the space dimension)

source

Numa.NodesArray — Type.

Array of nodes for a give polytope and order

source

Numa.Point — Type.

Type representing a point of D dimensions

source

Numa.PolynomialBasis — Method.

Create 1-dim polynomial basis of type [Lagrangian, Monomial] for a given order and a set of nodes. The nodes can be equispaced or Chebyshev nodes of second kind and take values [Equispaced, Chebyshev]

source

Numa.Polytope — Type.

Aggregation of all n-faces that compose the polytope boundary and the polytope itself, the classification of n-faces with respect to their dimension and type

source

Numa.Quadrature — Type.

Abstract type representing a quadrature rule on a Polytope in a space of D dimensions

source

Numa.TensorProductQuadrature — Type.

Tensor product quadrature rule (nodes and weights) on a hyper cube [-1,1]^D

source

Numa.TensorProductQuadratureOld — Type.

Tensor product quadrature rule (nodes and weights) integrating exactly 2´order´-1 polynomials

source

Numa.VectorValue — Type.

Type representing a vector value of D components

source

Numa.CellBasisValuesFromSingleInterpolation — Type.

Concrete implementation for the case of the same interpolation on all cells, but arbitrary sampling points in each cell. This is typically needed for unfitted methods

source

Numa.CellBasisWithMappedDerivaties — Type.

This type implements the result of mapderivatives

source

Numa.CellFieldValuesFromInterpolation — Type.

General implementation Accepts any kind of CellBasisValues{TB} representing the evaluated shape functions and CellFieldValues{TN} representing the "nodal" values of the field to be interpolated at the points, where the shape functions are evaluated. This implementation only assumes that outer(·,·) is defined for instances of TB and TN. The result is of type T

source

Numa.ConstantCellBasisValues — Type.

Concrete implementation for the case of the same interpolation and the same sampling points on all cells

source

Numa.GradOfCellBasisMappedWithJacob — Type.

Un-evaluated version of GradOfCellBasisValuesMappedWithJacob

source

Numa.GradOfCellBasisValuesMappedWithJacob — Type.

Values of the gradients of a cell basis mapped with a given Jacobian J stands for the type of the values of the Jacobian, i.e., gradient(Point{D},Val(D))

source

Numa.RefFE — Type.

Abstract Reference Finite Element

source