Gridap.Algebra

Gridap.Algebra — Module

The exported names are

AffineOperator

BackslashSolver

LUSolver

LinearSolver

NLSolver

NewtonRaphsonSolver

NonlinearOperator

NonlinearSolver

NumericalSetup

SparseMatrixCSR

SymSparseMatrixCSR

SymbolicSetup

add_entries!

add_entry!

allocate_coo_vectors

allocate_in_domain

allocate_in_range

allocate_jacobian

allocate_matrix

allocate_matrix_and_vector

allocate_residual

allocate_residual_and_jacobian

allocate_vector

colvals

copy_entries!

create_coo_vectors

fill_entries!

finalize_coo!

get_matrix

get_vector

getindices

getptr

hascolmajororder

hasrowmajororder

is_entry_stored

jacobian

jacobian!

muladd!

numerical_setup

numerical_setup!

push_coo!

residual

residual!

residual_and_jacobian

residual_and_jacobian!

scale_entries!

solve

solve!

sparse_from_coo

sparsecsr

symbolic_setup

symsparsecsr

test_linear_solver

test_nonlinear_operator

test_nonlinear_solver

zero_initial_guess

Gridap.Algebra.AffineOperator — Type

struct AffineOperator{A<:AbstractMatrix,B<:AbstractVector} <: NonlinearOperator
  matrix::A
  vector::B
end

Gridap.Algebra.BackslashSolver — Type

struct BackslashSolver <: LinearSolver end

Wrapper of the backslash solver available in julia This is typically faster than LU for a single solve

Gridap.Algebra.LUSolver — Type

struct LUSolver <: LinearSolver end

Wrapper of the LU solver available in julia

Gridap.Algebra.LinearSolver — Type

abstract type LinearSolver <: NonlinearSolver end

symbolic_setup(::LinearSolver,mat::AbstractMatrix)

test_linear_solver

Gridap.Algebra.NLSolver — Type

struct NLSolver <: NonlinearSolver
  # private fields
end

The cache generated when using this solver has a field result that hosts the result object generated by the underlying nlsolve function. It corresponds to the most latest solve.

Gridap.Algebra.NLSolver — Method

NLSolver(ls::LinearSolver;kwargs...)
NLSolver(;kwargs...)

Same kwargs as in nlsolve. If ls is provided, it is not possible to use the linsolve kw-argument.

Gridap.Algebra.NewtonRaphsonSolver — Type

struct NewtonRaphsonSolver <:NonlinearSolver
  # Private fields
end

Vanilla Newton-Raphson method

Gridap.Algebra.NonlinearOperator — Type

abstract type NonlinearOperator <: GridapType end

residual!(b::AbstractVector,op::NonlinearOperator,x::AbstractVector)

jacobian!(A::AbstractMatrix,op::NonlinearOperator,x::AbstractVector)

zero_initial_guess(op::NonlinearOperator)

allocate_residual(op::NonlinearOperator,x::AbstractVector)

allocate_jacobian(op::NonlinearOperator,x::AbstractVector)

Gridap.Algebra.NonlinearSolver — Type

abstract type NonlinearSolver <: GridapType end

solve!(x::AbstractVector,nls::NonlinearSolver,op::NonlinearOperator)

solve!(x::AbstractVector,nls::NonlinearSolver,op::NonlinearOperator, cache)

Gridap.Algebra.NumericalSetup — Type

abstract type NumericalSetup <: GridapType end

numerical_setup!(::NumericalSetup,mat::AbstractMatrix)

solve!(x::AbstractVector,::NumericalSetup,b::AbstractVector)

Gridap.Algebra.SparseMatrixCSR — Type

struct SparseMatrixCSR{Bi,Tv,Ti<:Integer} <: AbstractSparseMatrix{Tv,Ti}

Matrix type for storing Bi-based sparse matrices in the Compressed Sparse Row format. The standard way of constructing SparseMatrixCSR is through the sparsecsr function.

Gridap.Algebra.SymSparseMatrixCSR — Type

struct SymSparseMatrixCSR{T,Ti<:Integer} <: AbstractSparseMatrix{T,Ti}

Matrix type for storing symmetric sparse matrices in the Compressed Sparse Row format. The standard way of constructing SparseMatrixCSR is through the symsparsecsr function.

Gridap.Algebra.SymbolicSetup — Type

abstract type SymbolicSetup <: GridapType end

numerical_setup(::SymbolicSetup,mat::AbstractMatrix)

Gridap.Algebra.add_entries! — Function

add_entries!(a,b,combine=+)

Perform the operation combine element-wise in the entries of arrays a and b and store the result in array a. Returns a.

Gridap.Algebra.add_entry! — Function

add_entry!(A,v,i,combine=+)

Add an entry given its position and the operation to perform.

Gridap.Algebra.add_entry! — Function

add_entry!(A,v,i,j,combine=+)

Add an entry given its position and the operation to perform.

Gridap.Algebra.allocate_coo_vectors — Method

Gridap.Algebra.allocate_in_domain — Method

allocate_in_domain(::Type{V},matrix) where V

Allocate a vector of type V in the domain of matrix matrix.

Gridap.Algebra.allocate_in_range — Method

allocate_in_range(::Type{V},matrix) where V

Allocate a vector of type V in the range of matrix matrix.

Gridap.Algebra.allocate_jacobian — Method

allocate_jacobian(op::NonlinearOperator,x::AbstractVector)

Gridap.Algebra.allocate_residual — Method

allocate_residual(op::NonlinearOperator,x::AbstractVector)

Gridap.Algebra.allocate_residual_and_jacobian — Method

allocate_residual_and_jacobian(op::NonlinearOperator,x::AbstractVector)

Gridap.Algebra.allocate_vector — Method

allocate_vector(::Type{V},indices) where V

Allocate a vector of type V indexable at the indices indices

Gridap.Algebra.colvals — Method

colvals(S::AbstractSparseMatrix)

Return columns indices or raises an error (Depending on the SparseMatrix type)

Gridap.Algebra.copy_entries! — Method

copy_entries!(a,b)

Copy the entries of array b into array a. Returns a.

Gridap.Algebra.fill_entries! — Method

fill_entries!(a,v)

Fill the entries of array a with the value v. Returns a.

Gridap.Algebra.fill_entries! — Method

fill_entries!(A::AbstractSparseMatrix,v)

Fills the non-zero entries in the sparse matrix A with a given value v.

Gridap.Algebra.finalize_coo! — Method

finalize_coo!(::Type{T} where T,I,J,V,m,n)

Check and insert diagonal entries in COO vectors if needed.

Gridap.Algebra.get_matrix — Method

get_matrix(operator)

Return the matrix corresponding to the assembled left hand side of the operator. This matrix incorporates all boundary conditions and constraints.

Gridap.Algebra.get_vector — Method

get_vector(operator)

Return the vector corresponding to the assembled right hand side of the operator. This vector includes all boundary conditions and constraints.

Gridap.Algebra.getindices — Method

getindices(S::AbstractSparseMatrix)

Return row indices.

Gridap.Algebra.getptr — Method

getptr(S::AbstractSparseMatrix)

Return columns pointer.

Gridap.Algebra.hascolmajororder — Method

hascolmajororder(::Type{AbstractSparseMatrix})

Check if values are stored in col-major order. Return true.

Gridap.Algebra.hasrowmajororder — Method

hasrowmajororder(::Type{AbstractSparseMatrix})

Check if values are stored in row-major order. Return false.

Gridap.Algebra.is_entry_stored — Method

is_entry_stored(::Type{T} where T,i,j) -> Bool

Tells if the entry with coordinates [i,j] will be stored when calling function push_coo!

Gridap.Algebra.jacobian! — Method

jacobian!(A::AbstractMatrix,op::NonlinearOperator,x::AbstractVector)

Gridap.Algebra.jacobian — Method

jacobian(op::NonlinearOperator,x::AbstractVector)

Gridap.Algebra.muladd! — Method

muladd!(c,a,b)

Matrix multiply a*b and add to result to c. Returns c.

Gridap.Algebra.numerical_setup! — Method

numerical_setup!(::NumericalSetup,mat::AbstractMatrix)

Gridap.Algebra.numerical_setup — Method

numerical_setup(::SymbolicSetup,mat::AbstractMatrix) -> NumericalSetup

Gridap.Algebra.push_coo! — Method

push_coo!(::Type{T} where T, I,J,V,ik,jk,vk)

Inserts entries in COO vectors for further building a sparse matrix of type T.

Gridap.Algebra.residual! — Method

residual!(b::AbstractVector,op::NonlinearOperator,x::AbstractVector)

Gridap.Algebra.residual — Method

residual(op::NonlinearOperator,x::AbstractVector)

Gridap.Algebra.residual_and_jacobian! — Method

residual_and_jacobian!(
  b::AbstractVector, A::AbstractMatrix,
  op::NonlinearOperator, x::AbstractVector)

Gridap.Algebra.residual_and_jacobian — Method

residual_and_jacobian(op::NonlinearOperator,x::AbstractVector)

Gridap.Algebra.scale_entries! — Method

scale_entries!(a,v)

Scale the entries of array a with the value v. Returns a.

Gridap.Algebra.solve! — Method

solve!(
  x::AbstractVector,
  ls::LinearSolver,
  op::AffineOperator,
  cache,
  newmatrix::Bool)

Gridap.Algebra.solve! — Method

solve!(x::AbstractVector,nls::NonlinearSolver,op::NonlinearOperator,cache)

Solve using the cache object from a previous solve.

Gridap.Algebra.solve! — Method

solve!(x::AbstractVector,nls::NonlinearSolver,op::NonlinearOperator)

Usage:

cache = solve!(x,nls,op)

The returned cache object can be used in subsequent solves:

cache = solve!(x,nls,op,cache)

Gridap.Algebra.solve! — Method

solve!(x::AbstractVector,::NumericalSetup,b::AbstractVector)

Gridap.Algebra.solve — Method

solve(ls::LinearSolver,A::AbstractMatrix,b::AbstractVector)

Gridap.Algebra.solve — Method

solve(nls::NonlinearSolver,op::NonlinearOperator)

Creates and uses a zero initial guess.

Gridap.Algebra.sparse_from_coo — Method

sparse_from_coo(::Type{T} where T,I,J,V,m,n)

Gridap.Algebra.sparsecsr — Method

sparsecsr(args...)

Same args than sparse.

Gridap.Algebra.symbolic_setup — Method

symbolic_setup(::LinearSolver,mat::AbstractMatrix) -> SymbolicSetup

Gridap.Algebra.symsparsecsr — Method

symsparsecsr(args...)

Same args than sparse.

Gridap.Algebra.test_linear_solver — Method

test_linear_solver(
  ls::LinearSolver,
  A::AbstractMatrix,
  b::AbstractVector,
  x::AbstractVector)

Gridap.Algebra.test_nonlinear_operator — Function

test_nonlinear_operator(
  op::NonlinearOperator,
  x::AbstractVector,
  b::AbstractVector,
  pred=isapprox;
  jac=nothing)

Gridap.Algebra.test_nonlinear_solver — Function

test_nonlinear_solver(
  nls::NonlinearSolver,
  op::NonlinearOperator,
  x0::AbstractVector,
  x::AbstractVector,
  pred::Function=isapprox)

Gridap.Algebra.zero_initial_guess — Method

zero_initial_guess(op::NonlinearOperator)

Base.convert — Method

convert(::Type{AbstractSparseMatrix}, x::AbstractSparseMatrix)

Convert x to a value of the first type given.

Base.count — Method

count(pred, S::AbstractSparseMatrix) -> Integer

Count the number of elements in nonzeros(S) for which predicate pred returns true.

SparseArrays.findnz — Method

findnz(S::AbstractSparseMatrix)

Return a tuple (I, J, V) where I and J are the row and column indices of the stored ("structurally non-zero") values in sparse matrix A, and V is a vector of the values.

SparseArrays.nnz — Method

nnz(S::AbstractSparseMatrix)

Returns the number of stored (filled) elements in a sparse array.

SparseArrays.nonzeros — Method

nonzeros(S::AbstractSparseMatrix)

Return a vector of the structural nonzero values in sparse array S. This includes zeros that are explicitly stored in the sparse array. The returned vector points directly to the internal nonzero storage of S, and any modifications to the returned vector will mutate S as well.

SparseArrays.nzrange — Method

nzrange(S::AbstractSparseMatrix, index::Integer) where {Bi}

Return the range of indices to the structural nonzero values of a sparse matrix index (Row or column depending on the compression type).

SparseArrays.rowvals — Method

rowvals(S::AbstractSparseMatrix)

Return row indices or raises an error (Depending on the SparseMatrix type)