Gridap.FESpaces

Gridap.FESpaces — Module

The exported names are

AffineFEOperator

Assembler

AssemblyStrategy

BasisStyle

CLagrangianFESpace

CellConformity

CellFE

ConstantFESpace

Constrained

ConstraintStyle

DefaultAssemblyStrategy

DirichletFESpace

DiscreteModelWithFEMap

EvaluationFunction

FEBasis

FEFunction

FEOperator

FESolver

FESpace

FESpaceWithConstantFixed

FESpaceWithLinearConstraints

FiniteElements

GenericAssemblyStrategy

GenericSparseMatrixAssembler

GridWithFEMap

HomogeneousTrialFESpace

HomogeneousTrialFESpace!

LinearFESolver

LocalOperator

NonlinearFESolver

PatchAssembler

SingleFieldFEFunction

SingleFieldFESpace

SparseMatrixAssembler

TestFESpace

TrialFESpace

TrialFESpace!

UnConstrained

UnconstrainedFESpace

ZeroMeanFESpace

add_mesh_displacement!

allocate_matrix

allocate_matrix_and_vector

allocate_vector

assemble_matrix

assemble_matrix!

assemble_matrix_add!

assemble_matrix_and_vector

assemble_matrix_and_vector!

assemble_matrix_and_vector_add!

assemble_vector

assemble_vector!

assemble_vector_add!

col_map

col_mask

collect_cell_matrix

collect_cell_matrix_and_vector

collect_cell_vector

compute_cell_space

compute_conforming_cell_dofs

compute_dirichlet_values_for_tags

compute_dirichlet_values_for_tags!

gather_dirichlet_values

gather_dirichlet_values!

gather_free_and_dirichlet_values!

gather_free_values

gather_free_values!

get_algebraic_operator

get_cell_constraints

get_cell_dof_ids

get_cell_dof_values

get_cell_is_dirichlet

get_cell_isconstrained

get_cols

get_dirichlet_dof_ids

get_dirichlet_dof_tag

get_dirichlet_dof_values

get_dirichlet_values

get_dof_value_type

get_fe_basis

get_fe_dof_basis

get_fe_space

get_free_dof_ids

get_free_dof_values

get_free_values

get_matrix_builder

get_matrix_type

get_rows

get_test

get_trial

get_trial_fe_basis

get_vector_builder

get_vector_type

has_constraints

interpolate

interpolate!

interpolate_dirichlet

interpolate_dirichlet!

interpolate_everywhere

interpolate_everywhere!

num_dirichlet_dofs

num_dirichlet_tags

num_free_dofs

numeric_loop_matrix!

numeric_loop_matrix_and_vector!

numeric_loop_vector!

row_map

row_mask

scatter_free_and_dirichlet_values

symbolic_loop_matrix!

symbolic_loop_matrix_and_vector!

symbolic_loop_vector!

test_assembler

test_fe_function

test_fe_operator

test_fe_solver

test_fe_space

test_single_field_fe_space

test_sparse_matrix_assembler

update_coordinates!

zero_dirichlet_values

zero_free_values

Gridap.FESpaces.AffineFEOperator — Type

AffineFEOperator

Represent a fully assembled affine (linear) finite element problem. See also FEOperator

Gridap.FESpaces.AffineFEOperator — Method

Gridap.FESpaces.Assembler — Type

Gridap.FESpaces.AssemblyStrategy — Type

Gridap.FESpaces.CellConformity — Type

Minimum data required to describe dof ownership. At this moment, the cell-wise ownership is compressed on cell types. This can be relaxed in the future, to have an arbitrary cell-wise dof ownership.

Gridap.FESpaces.CellConformity — Method

Generate A CellConformity from a vector of reference fes

Gridap.FESpaces.CellFE — Type

Minimum data required to build a conforming FE space. At this moment, the some cell-wise info is compressed on cell types. This can be relaxed in the future, and have an arbitrary cell-wise data.

Gridap.FESpaces.CellFE — Method

Generate a CellFE from a vector of reference fes

Gridap.FESpaces.ConstantFESpace — Type

struct ConstantFESpace <: SingleFieldFESpace
  # private fields
end

ConstantFESpace(model::DiscreteModel; vector_type=Vector{Float64}, field_type=Float64)
ConstantFESpace(trian::Triangulation; vector_type=Vector{Float64}, field_type=Float64)

FESpace that is constant over the provided model/triangulation. Typically used as lagrange multipliers. The kwargs vector_type and field_type are used to specify the types of the dof-vector and dof-value respectively.

Gridap.FESpaces.DirichletFESpace — Type

struct DirichletFESpace <: SingleFieldFESpace
  space::SingleFieldFESpace
end

Gridap.FESpaces.FEFunction — Method

Gridap.FESpaces.FEFunction — Method

FEFunction(
  fs::SingleFieldFESpace, free_values::AbstractVector, dirichlet_values::AbstractVector)

The resulting FEFunction will be in the space if and only if dirichlet_values are the ones provided by get_dirichlet_dof_values(fs)

Gridap.FESpaces.FEOperator — Type

abstract type FEOperator <: GridapType

A FEOperator contains finite element problem, that is assembled as far as possible and ready to be solved. See also FETerm

Gridap.FESpaces.FESolver — Type

Gridap.FESpaces.FESolver — Method

Gridap.FESpaces.FESpaceWithConstantFixed — Type

FESpaceWithConstantFixed(space::SingleFieldFESpace, fix_constant::Bool,
dof_to_fix::Int=num_free_dofs(space))

Gridap.FESpaces.GridWithFEMap — Type

Given a Discrete Model and a reffe, builds a new grid in which the geometrical map is a FEFunction. This is useful when considering geometrical maps that are the result of a FE problem (mesh displacement).

Gridap.FESpaces.LinearFESolver — Type

The solver that solves a LinearFEOperator

Gridap.FESpaces.LinearFESolver — Method

Gridap.FESpaces.LocalOperator — Type

struct LocalOperator end

Gridap.FESpaces.MultiConstantFESpace — Type

struct MultiConstantFESpace{N,V,T} <: SingleFieldFESpace end

MultiConstantFESpace(model::DiscreteModel,tags::Vector,D::Integer)
MultiConstantFESpace(trians::Vector{<:BoundaryTriangulation{Df}})

Extension of ConstantFESpace, representing a FESpace which is constant on each of it's N triangulations.

Gridap.FESpaces.NodeToDofGlue — Type

struct NodeToDofGlue{T}
  free_dof_to_node::Vector{Int32}
  free_dof_to_comp::Vector{Int16}
  dirichlet_dof_to_node::Vector{Int32}
  dirichlet_dof_to_comp::Vector{Int16}
  node_and_comp_to_dof::Vector{T}
end

Gridap.FESpaces.NonlinearFESolver — Type

A general NonlinearFESolver

Gridap.FESpaces.NonlinearFESolver — Method

Gridap.FESpaces.SingleFieldFESpace — Type

Gridap.FESpaces.SparseMatrixAssembler — Type

Gridap.FESpaces.SparseMatrixAssembler — Method

Gridap.FESpaces.TrialFESpace — Method

Gridap.FESpaces.TrialFESpace — Method

Gridap.FESpaces.UnconstrainedFESpace — Type

Generic implementation of an unconstrained single-field FE space Private fields and type parameters

Gridap.FESpaces.ZeroMeanFESpace — Type

struct ZeroMeanFESpace <: SingleFieldFESpace
  # private fields
end

Gridap.FESpaces.ZeroMeanFESpace — Method

Base.zero — Method

Gridap.Algebra.allocate_jacobian — Method

Gridap.Algebra.allocate_matrix — Method

Gridap.Algebra.allocate_matrix_and_vector — Method

Gridap.Algebra.allocate_residual — Method

Gridap.Algebra.allocate_vector — Method

Gridap.Algebra.get_matrix — Method

Gridap.Algebra.get_vector — Method

Gridap.Algebra.jacobian! — Method

jacobian!(A, op, u)

Inplace version of jacobian.

Gridap.Algebra.jacobian — Method

jacobian(op, u)

Compute the jacobian of an operator op. See also get_algebraic_operator, residual_and_jacobian!.

Gridap.Algebra.residual! — Method

residual!(b, op, u)

Inplace version of residual.

Gridap.Algebra.residual — Method

residual(op, u)

Compute the residual of op at u. See also residual_and_jacobian

Gridap.Algebra.residual_and_jacobian! — Method

residual_and_jacobian!(b, A, op, u)

Inplace version of residual_and_jacobian.

Gridap.Algebra.residual_and_jacobian — Method

residual, jacobian =

residual_and_jacobian(op, u)

Compute the residual and jacobian of an operator op at a given point u. Depending on the nature of op the point u can either be a plain array or a FEFunction.

See also jacobian, residual, get_algebraic_operator.

Gridap.Algebra.solve! — Method

uh, cache = solve!(uh,solver,op,cache)

This function changes the state of the input and can render it in a corrupted state. It is recommended to rewrite the input uh with the output as illustrated to prevent any issue. If cache===nothing, then it creates a new cache object.

Gridap.Algebra.solve! — Method

uh, cache = solve!(uh,solver,op)

This function changes the state of the input and can render it in a corrupted state. It is recommended to rewrite the input uh with the output as illustrated to prevent any issue.

Gridap.Algebra.solve — Method

Gridap.Algebra.solve — Method

Solve that allocates, and sets initial guess to zero and returns the solution

Gridap.FESpaces.CLagrangianFESpace — Method

CLagrangianFESpace(
::Type{T},
grid::Triangulation,
vector_type::Type,
node_to_tag::AbstractVector,
tag_to_mask::AbstractVector) where T

Gridap.FESpaces.CLagrangianFESpace — Method

CLagrangianFESpace(::Type{T},grid::Triangulation) where T

Gridap.FESpaces.TestFESpace — Method

Gridap.FESpaces.TrialFESpace! — Method

Gridap.FESpaces.TrialFESpace! — Method

Gridap.FESpaces.assemble_matrix! — Method

Gridap.FESpaces.assemble_matrix — Method

Gridap.FESpaces.assemble_matrix_add! — Method

Gridap.FESpaces.assemble_matrix_and_vector! — Method

Gridap.FESpaces.assemble_matrix_and_vector — Method

Gridap.FESpaces.assemble_vector! — Method

Gridap.FESpaces.assemble_vector — Method

Gridap.FESpaces.assemble_vector_add! — Method

Gridap.FESpaces.col_map — Method

Gridap.FESpaces.col_mask — Method

Gridap.FESpaces.compute_conforming_cell_dofs — Function

The result is the tuple

(cell_dofs, nfree, ndiri, dirichlet_dof_tag, dirichlet_cells)

If dirichlet_components is given, then get_dof_to_comp has to be defined for the reference elements in reffes.

Gridap.FESpaces.compute_dirichlet_values_for_tags — Method

Gridap.FESpaces.gather_dirichlet_values! — Method

Gridap.FESpaces.gather_dirichlet_values — Method

Gridap.FESpaces.gather_free_and_dirichlet_values! — Method

Gridap.FESpaces.gather_free_and_dirichlet_values — Method

Gridap.FESpaces.gather_free_values! — Method

Gridap.FESpaces.gather_free_values — Method

Gridap.FESpaces.generate_cell_dof_mask — Method

generate_cell_dof_mask(
  scell_conformity::CellConformity,
  tcell_to_scell::Vector{<:Integer},
  d_to_tcell_to_tdface,
  d_to_tdface_to_mask;
  reverse::Bool=false
)

Given a CellConformity object, defined on a set of source cells (scell), and given a cell/face/edge/node masks on a set of target cells (tcell), this function generates a mask for the degrees of freedom (dofs) on the target cells.

Parameters:

scell_conformity: The CellConformity object on the source cells (scell).

tcell_to_scell: A vector mapping target cells (tcell) to source cells (scell).

d_to_tcell_to_tdface: For each dimension d, a Table mapping target cells to its d-faces (edges, faces, ...)

d_to_tdface_to_mask: For each dimension d, an array of booleans indicating whether the d-face is masked or not.

Returns:

tcell_dof_mask: A vector that for each target cell contains a boolean vector of size equal to the number of dofs in that cell, containing the mask.

Modes of operation:

If reverse = false, the function generates a mask where the dofs are true if the corresponding d-face is masked (true).

If reverse = true, the mask is reversed, meaning that the dofs are true if the corresponding d-face is not masked (false).

Gridap.FESpaces.get_algebraic_operator — Method

get_algebraic_operator(feop)

Return an "algebraic view" of an operator. Algebraic means, that the resulting operator acts on plain arrays, instead of FEFunctions. This can be useful for solving with external tools like NLsolve.jl. See also FEOperator.

Gridap.FESpaces.get_assembly_strategy — Method

Gridap.FESpaces.get_cell_conformity — Method

Gridap.FESpaces.get_cell_dof_ids — Method

Gridap.FESpaces.get_cell_dof_values — Method

Gridap.FESpaces.get_cols — Method

Gridap.FESpaces.get_dirichlet_dof_ids — Method

Gridap.FESpaces.get_dirichlet_dof_tag — Method

Gridap.FESpaces.get_dirichlet_dof_values — Method

Gridap.FESpaces.get_fe_basis — Method

Gridap.FESpaces.get_fe_dof_basis — Method

Gridap.FESpaces.get_fe_space — Method

Gridap.FESpaces.get_free_dof_ids — Method

Gridap.FESpaces.get_free_dof_values — Method

Gridap.FESpaces.get_matrix_builder — Method

Gridap.FESpaces.get_rows — Method

Gridap.FESpaces.get_test — Method

Gridap.FESpaces.get_trial — Method

Gridap.FESpaces.get_vector_builder — Method

Gridap.FESpaces.interpolate! — Method

Gridap.FESpaces.interpolate — Method

The resulting FE function is in the space (in particular it fulfills Dirichlet BCs even in the case that the given cell field does not fulfill them)

Gridap.FESpaces.interpolate_dirichlet! — Method

Gridap.FESpaces.interpolate_dirichlet — Method

Gridap.FESpaces.interpolate_everywhere! — Method

Gridap.FESpaces.interpolate_everywhere — Method

like interpolate, but also compute new degrees of freedom for the dirichlet component. The resulting FEFunction does not necessary belongs to the underlying space

Gridap.FESpaces.num_dirichlet_dofs — Method

Gridap.FESpaces.num_dirichlet_tags — Method

Gridap.FESpaces.num_free_dofs — Method

Gridap.FESpaces.row_map — Method

Gridap.FESpaces.row_mask — Method

Gridap.FESpaces.scatter_free_and_dirichlet_values — Method

Gridap.FESpaces.test_assembler — Method

Gridap.FESpaces.test_fe_function — Method

Gridap.FESpaces.test_fe_operator — Method

Gridap.FESpaces.test_fe_solver — Function

Gridap.FESpaces.test_fe_space — Method

Gridap.FESpaces.test_single_field_fe_space — Function

Gridap.FESpaces.zero_dirichlet_values — Method

Gridap.FESpaces.zero_free_values — Method