Gridap.FESpaces

Gridap.FESpaces — Module

The exported names are

AffineFEOperator

AffineFETerm

AffineFETermFromCellMatVec

Assembler

AssemblyStrategy

CLagrangianFESpace

CellBasis

CellDofBasis

CellField

CellMatrixField

CurlConformingFESpace

DefaultAssemblyStrategy

DirichletFESpace

DiscontinuousFESpace

DivConformingFESpace

ExtendedFESpace

FECellBasisStyle

FEFunction

FEFunctionStyle

FEOperator

FEOperatorFromTerms

FESolver

FESource

FESpace

FESpaceWithLastDofRemoved

FESpaceWithLinearConstraints

FETerm

FETermFromCellJacRes

GenericCellBasis

GenericCellDofBasis

GenericCellMatrixField

GradConformingFESpace

GridapType

HomogeneousTrialFESpace

HomogeneousTrialFESpace!

LinearFESolver

LinearFETerm

NonlinearFESolver

SingleFieldFEFunction

SingleFieldFESpace

SparseMatrixAssembler

TestFESpace

TrialFESpace

TrialFESpace!

TrialStyle

UnconstrainedFESpace

ZeroMeanFESpace

allocate_matrix

allocate_matrix_and_vector

allocate_vector

apply_cellmatrix

apply_cellmatvec

apply_cellvector

apply_constraints_matrix_and_vector_cols

apply_constraints_matrix_and_vector_rows

apply_constraints_matrix_cols

apply_constraints_matrix_rows

apply_constraints_vector

apply_statelaw

assemble_matrix

assemble_matrix!

assemble_matrix_add!

assemble_matrix_and_vector

assemble_matrix_and_vector!

assemble_matrix_and_vector_add!

assemble_vector

assemble_vector!

assemble_vector_add!

attach_dirichlet_bcs

col_map

col_mask

collect_cell_jacobian

collect_cell_jacobian_and_residual

collect_cell_matrix

collect_cell_matrix_and_vector

collect_cell_residual

collect_cell_vector

compute_conforming_cell_dofs

compute_dirichlet_values

compute_dirichlet_values_for_tags

compute_dirichlet_values_for_tags!

compute_free_and_dirichlet_values

compute_free_values

constraint_style

count_matrix_and_vector_nnz_coo

count_matrix_nnz_coo

fill_matrix_and_vector_coo_numeric!

fill_matrix_and_vector_coo_symbolic!

fill_matrix_coo_numeric!

fill_matrix_coo_symbolic!

gather_dirichlet_values

gather_free_and_dirichlet_values

gather_free_values

get_algebraic_operator

get_cell_basis

get_cell_constraints

get_cell_dof_basis

get_cell_dofs

get_cell_isconstrained

get_cell_jacobian

get_cell_jacobian_and_residual

get_cell_matrix

get_cell_residual

get_cell_values

get_cell_vector

get_constraint_kernel_matrix_cols

get_constraint_kernel_matrix_rows

get_constraint_kernel_vector

get_dirichlet_dof_tag

get_dirichlet_values

get_fe_space

get_free_values

get_matrix_type

get_test

get_trial

get_vector_type

has_constraints

interpolate

interpolate_dirichlet

interpolate_everywhere

is_a_fe_cell_basis

is_a_fe_function

is_test

is_trial

@law

num_dirichlet_dofs

num_dirichlet_tags

num_free_dofs

operate

row_map

row_mask

scatter_free_and_dirichlet_values

@statelaw

test_assembler

test_cell_basis

test_cell_dof_basis

test_cell_matrix_field

test_fe_function

test_fe_operator

test_fe_solver

test_fe_space

test_single_field_fe_space

test_sparse_matrix_assembler

update_state_variables!

zero_dirichlet_values

zero_free_values

Gridap.FESpaces.AffineFEOperator — Type

Gridap.FESpaces.AffineFEOperator — Method

Gridap.FESpaces.AffineFEOperator — Method

AffineFEOperator(test::FESpace,trial::FESpace,assem::Assembler,terms::AffineFETerm...) AffineFEOperator(test::FESpace,trial::FESpace,terms::AffineFETerm...)

Gridap.FESpaces.AffineFETerm — Type

Gridap.FESpaces.AffineFETerm — Method

Gridap.FESpaces.AffineFETermFromCellMatVec — Type

Gridap.FESpaces.Assembler — Type

Gridap.FESpaces.AssemblyStrategy — Type

Gridap.FESpaces.CLagrangianFESpace — Type

struct CLagrangianFESpace{S} <: SingleFieldFESpace
  grid::Grid
  dof_to_node::Vector{Int}
  dof_to_comp::Vector{Int8}
  node_and_comp_to_dof::Vector{S}
  # + private fields
end

Gridap.FESpaces.CLagrangianFESpace — Method

CLagrangianFESpace(::Type{T},grid::Grid) where T

Gridap.FESpaces.CellBasis — Type

Gridap.FESpaces.CellDofBasis — Type

abstract type CellDofBasis end

Abstract type that represents a cell array of Dof. The main motivation for its definition is to provide a trait that informs whether the Dof entries are defined for functions in the reference or physical space

Gridap.FESpaces.CellMatrixField — Type

Gridap.FESpaces.DirichletFESpace — Type

struct DirichletFESpace <: SingleFieldFESpace
  space::SingleFieldFESpace
end

Gridap.FESpaces.ExtendedFESpace — Type

Gridap.FESpaces.FEOperator — Type

Gridap.FESpaces.FEOperator — Method

Gridap.FESpaces.FEOperator — Method

Gridap.FESpaces.FESolver — Type

Gridap.FESpaces.FESolver — Method

Gridap.FESpaces.FESource — Type

Gridap.FESpaces.FESource — Method

Gridap.FESpaces.FESpace — Type

Gridap.FESpaces.FESpace — Method

Gridap.FESpaces.FESpaceWithLastDofRemoved — Type

struct FESpaceWithLastDofRemoved <: SingleFieldFESpace
  space::SingleFieldFESpace
end

Gridap.FESpaces.FESpaceWithLastDofRemoved — Method

FESpaceWithLastDofRemoved(space::SingleFieldFESpace)

Gridap.FESpaces.FETerm — Type

Gridap.FESpaces.FETerm — Method

Gridap.FESpaces.FETermFromCellJacRes — Type

Gridap.FESpaces.GenericCellBasis — Type

Gridap.FESpaces.GenericCellDofBasis — Type

Gridap.FESpaces.GenericCellMatrixField — Type

Gridap.FESpaces.LinearFESolver — Type

The solver that solves a LinearFEOperator

Gridap.FESpaces.LinearFESolver — Method

Gridap.FESpaces.LinearFETerm — Type

Gridap.FESpaces.LinearFETerm — Method

Gridap.FESpaces.NonlinearFESolver — Type

A general NonlinearFESolver

Gridap.FESpaces.NonlinearFESolver — Method

Gridap.FESpaces.SingleFieldFEFunction — Type

Gridap.FESpaces.SingleFieldFEFunction — Method

Gridap.FESpaces.SingleFieldFESpace — Type

Gridap.FESpaces.SparseMatrixAssembler — Type

Gridap.FESpaces.SparseMatrixAssembler — Method

Gridap.FESpaces.TrialFESpace — Method

Gridap.FESpaces.TrialFESpace — Method

Gridap.FESpaces.UnconstrainedFESpace — Type

Generic implementation of an unconstrained single-field FE space Private fields and type parameters

Gridap.FESpaces.UnconstrainedFESpace — Method

Gridap.FESpaces.ZeroMeanFESpace — Type

struct ZeroMeanFESpace <: SingleFieldFESpace
  # private fields
end

Gridap.FESpaces.ZeroMeanFESpace — Method

ZeroMeanFESpace(
  space::SingleFieldFESpace,
  trian::Triangulation,
  quad::CellQuadrature)

Gridap.Algebra.allocate_matrix — Method

Gridap.Algebra.allocate_matrix_and_vector — Method

Gridap.Algebra.allocate_vector — Method

Gridap.FESpaces.CurlConformingFESpace — Method

CurlConformingFESpace(
  reffes::Vector{<:ReferenceFE},
  model::DiscreteModel,
  face_labeling::FaceLabeling,
  dirichlet_tags)

Gridap.FESpaces.DiscontinuousFESpace — Function

DiscontinuousFESpace(reffes::Vector{<:ReferenceFE}, trian::Triangulation)

Gridap.FESpaces.DivConformingFESpace — Method

DivConformingFESpace(
  reffes::Vector{<:ReferenceFE},
  model::DiscreteModel,
  face_labeling::FaceLabeling,
  dirichlet_tags)

Gridap.FESpaces.FECellBasisStyle — Method

Gridap.FESpaces.FEFunction — Method

Gridap.FESpaces.FEFunction — Method

FEFunction(
  fs::SingleFieldFESpace, free_values::AbstractVector, dirichlet_values::AbstractVector)

The resulting FEFunction will be in the space if and only if dirichlet_values are the ones provided by get_dirichlet_values(fs)

Gridap.FESpaces.FEFunctionStyle — Method

Types marked with this trait need to implement the following queries

[get_free_values(object)]

[get_fe_space(object)]

Gridap.FESpaces.GradConformingFESpace — Function

Gridap.FESpaces.GradConformingFESpace — Function

Gridap.FESpaces.TestFESpace — Method

Gridap.FESpaces.TrialFESpace! — Method

Gridap.FESpaces.TrialFESpace! — Method

Gridap.FESpaces.TrialStyle — Method

Gridap.FESpaces.apply_cellmatrix — Method

Gridap.FESpaces.apply_cellmatvec — Method

Gridap.FESpaces.apply_cellvector — Method

Gridap.FESpaces.apply_constraints_matrix_and_vector_cols — Method

Gridap.FESpaces.apply_constraints_matrix_and_vector_rows — Method

Gridap.FESpaces.apply_constraints_matrix_cols — Method

Gridap.FESpaces.apply_constraints_matrix_rows — Method

Gridap.FESpaces.apply_constraints_vector — Method

Gridap.FESpaces.apply_statelaw — Method

Gridap.FESpaces.assemble_matrix! — Method

Gridap.FESpaces.assemble_matrix — Method

Gridap.FESpaces.assemble_matrix_add! — Method

Gridap.FESpaces.assemble_matrix_and_vector! — Method

Gridap.FESpaces.assemble_matrix_and_vector — Method

Gridap.FESpaces.assemble_vector! — Method

Gridap.FESpaces.assemble_vector — Method

Gridap.FESpaces.assemble_vector_add! — Method

Gridap.FESpaces.attach_dirichlet_bcs — Method

Gridap.FESpaces.col_map — Method

Gridap.FESpaces.col_mask — Method

Gridap.FESpaces.collect_cell_jacobian — Method

Gridap.FESpaces.collect_cell_jacobian_and_residual — Method

Gridap.FESpaces.collect_cell_matrix — Method

Gridap.FESpaces.collect_cell_matrix_and_vector — Method

Gridap.FESpaces.collect_cell_residual — Method

Gridap.FESpaces.collect_cell_vector — Method

Gridap.FESpaces.compute_conforming_cell_dofs — Function

computeconformingcelldofs( reffes, gridtopology, facelabeling, dirichlettags)

computeconformingcelldofs( reffes, gridtopology, facelabeling, dirichlettags, dirichlet_components)

The result is the tuple

(cell_dofs, nfree, ndiri, dirichlet_dof_tag, dirichlet_cells)

Assumes that the reffes are aligned with the cell type in the grid_topology and that it is possible to build a conforming space without imposing constraints

If dirichlet_components is given, then get_dof_to_comp has to be defined for the reference elements in reffes.

Gridap.FESpaces.compute_dirichlet_values_for_tags — Method

Gridap.FESpaces.constraint_style — Method

Gridap.FESpaces.count_matrix_and_vector_nnz_coo — Method

Gridap.FESpaces.count_matrix_nnz_coo — Method

Gridap.FESpaces.fill_matrix_and_vector_coo_numeric! — Function

Gridap.FESpaces.fill_matrix_coo_numeric! — Function

Gridap.FESpaces.fill_matrix_coo_symbolic! — Function

Gridap.FESpaces.gather_dirichlet_values — Method

Gridap.FESpaces.gather_free_and_dirichlet_values — Method

Gridap.FESpaces.gather_free_values — Method

Gridap.FESpaces.get_algebraic_operator — Method

Gridap.FESpaces.get_cell_basis — Method

Gridap.FESpaces.get_cell_dof_basis — Method

Gridap.FESpaces.get_cell_dofs — Method

Gridap.FESpaces.get_cell_jacobian — Method

Returns an object representing the contribution to the Jacobian of the given term. Returns nothing if the term has not contribution to the Jacobian (typically for source terms)

Gridap.FESpaces.get_cell_jacobian_and_residual — Method

Gridap.FESpaces.get_cell_matrix — Method

Returns an object representing the contribution to the system matrix of the given term. Returns nothing if the term has not contribution (typically for source terms)

Gridap.FESpaces.get_cell_residual — Method

Returns an object representing the contribution to the residual of the given term. Returns always something.

Gridap.FESpaces.get_cell_values — Method

Gridap.FESpaces.get_cell_values — Method

Gridap.FESpaces.get_cell_vector — Method

Returns an object (e.g. a CellVector) representing the contribution to the system rhs of the given term. Returns nothing if the term has not contribution (typically for linear terms)

Gridap.FESpaces.get_constraint_kernel_matrix_cols — Method

Gridap.FESpaces.get_constraint_kernel_matrix_rows — Method

Gridap.FESpaces.get_constraint_kernel_vector — Method

Gridap.FESpaces.get_dirichlet_dof_tag — Method

Gridap.FESpaces.get_dirichlet_values — Method

Gridap.FESpaces.get_fe_space — Method

Gridap.FESpaces.get_free_values — Method

Gridap.FESpaces.get_matrix_type — Method

Gridap.FESpaces.get_test — Method

Gridap.FESpaces.get_test — Method

Gridap.FESpaces.get_trial — Method

Gridap.FESpaces.get_trial — Method

Gridap.FESpaces.get_vector_type — Method

Gridap.FESpaces.has_constraints — Method

Gridap.FESpaces.interpolate — Method

The resulting FE function is in the space (in particular it fulfills Dirichlet BCs even in the case that the given cell field does not fulfill them)

Gridap.FESpaces.interpolate_dirichlet — Method

Gridap.FESpaces.interpolate_everywhere — Method

like interpolate, but also compute new degrees of freedom for the dirichlet component. The resulting FEFunction does not necessary belongs to the underlying space

Gridap.FESpaces.is_a_fe_cell_basis — Method

Gridap.FESpaces.is_a_fe_function — Method

Gridap.FESpaces.is_test — Method

Gridap.FESpaces.is_trial — Method

Gridap.FESpaces.num_dirichlet_dofs — Method

Gridap.FESpaces.num_dirichlet_tags — Method

Gridap.FESpaces.num_free_dofs — Method

Gridap.FESpaces.row_map — Method

Gridap.FESpaces.row_mask — Method

Gridap.FESpaces.scatter_free_and_dirichlet_values — Method

Gridap.FESpaces.test_assembler — Method

Gridap.FESpaces.test_cell_basis — Method

Gridap.FESpaces.test_cell_matrix_field — Method

Gridap.FESpaces.test_fe_function — Method

Gridap.FESpaces.test_fe_operator — Method

Gridap.FESpaces.test_fe_solver — Function

Gridap.FESpaces.test_fe_space — Method

Gridap.FESpaces.test_single_field_fe_space — Function

Gridap.FESpaces.update_state_variables! — Method

Gridap.FESpaces.zero_dirichlet_values — Method

Gridap.FESpaces.zero_free_values — Method

Gridap.FESpaces.@law — Macro

Gridap.FESpaces.@statelaw — Macro

Base.zero — Method

Gridap.Algebra.allocate_jacobian — Method

Gridap.Algebra.allocate_residual — Method

Gridap.Algebra.get_matrix — Method

Gridap.Algebra.get_vector — Method

Gridap.Algebra.jacobian! — Method

Gridap.Algebra.residual! — Method

Gridap.Algebra.residual_and_jacobian! — Method

Gridap.Algebra.residual_and_jacobian — Method

Gridap.Algebra.solve! — Method

uh, cache = solve!(uh,solver,op,cache)

This function changes the state of the input and can render it in a corrupted state. It is recommended to rewrite the input uh with the output as illustrated to prevent any issue. If cache===nothing, then it creates a new cache object.

Gridap.Algebra.solve! — Method

uh, cache = solve!(uh,solver,op)

This function changes the state of the input and can render it in a corrupted state. It is recommended to rewrite the input uh with the output as illustrated to prevent any issue.

Gridap.Algebra.solve — Method

Gridap.Algebra.solve — Method

Solve that allocates, and sets initial guess to zero and returns the solution

Gridap.FESpaces.compute_cell_space — Method

Gridap.FESpaces.compute_cell_space_physical — Method

It creates the cell-wise DOF basis and shape functions, when the DOFs are evaluated at the physical space. The DOFs (moments) for the prebasis are assumed to be computable at a reference FE space.

Gridap.FESpaces.compute_discontinuous_cell_dofs — Method

Gridap.FESpaces.get_assembly_strategy — Method

Gridap.FESpaces.get_cell_matrix_and_vector — Method

Gridap.Fields.evaluate — Method

evaluate(dof_array::CellDofBasis,field_array::AbstractArray)

Evaluates the CellDofBasis for the Field objects at the array field element by element.

The result is numerically equivalent to

map(evaluate_dof, dof_array.array, field_array)

but it is described with a more memory-friendly lazy type.

Gridap.Geometry.get_cell_id — Method

Gridap.Geometry.get_cell_map — Method

Gridap.Geometry.get_cell_shapefuns — Method