Gridap.Geometry

Gridap.Geometry — Module

Exported names are

BoundaryTriangulation

CartesianDescriptor

CartesianDiscreteModel

CartesianGrid

CellField

CellQuadrature

DiscreteModel

DiscreteModelFromFile

DiscreteModelPortion

FaceLabeling

GenericBoundaryTriangulation

GenericCellField

Grid

GridPortion

GridTopology

InterfaceTriangulation

OrientationStyle

QPointCellField

RegularityStyle

RestrictedDiscreteModel

RestrictedTriangulation

SkeletonCellField

SkeletonPair

SkeletonTriangulation

Triangulation

TriangulationPortion

UnstructuredDiscreteModel

UnstructuredGrid

UnstructuredGridTopology

add_tag!

add_tag_from_tags!

compute_cell_faces

compute_cell_permutations

compute_face_nodes

compute_face_own_nodes

compute_face_vertices

compute_isboundary_face

compute_linear_grid

compute_node_face_owner

compute_reference_grid

compute_reffaces

compute_vertex_node

convert_to_cell_field

get_cartesian_descriptor

get_cell_coordinates

get_cell_entity

get_cell_faces

get_cell_id

get_cell_map

get_cell_nodes

get_cell_permutations

get_cell_reffes

get_cell_shapefuns

get_cell_type

get_cell_vertices

get_face_entity

get_face_labeling

get_face_mask

get_face_tag

get_face_tag_index

get_face_to_cell

get_face_to_cell_map

get_face_to_lface

get_grid

get_grid_topology

get_isboundary_face

get_left_boundary

get_node_face_owner

get_normal_vector

get_physical_coordinate

get_polytopes

get_reffaces_offsets

get_reffes

get_right_boundary

get_tag_entities

get_tag_from_name

get_tag_name

get_tags_from_names

get_triangulation

get_volume_triangulation

is_oriented

is_regular

jump

mean

num_cells

num_entities

num_tags

restrict

similar_object

test_boundary_triangulation

test_cell_field

test_discrete_model

test_grid

test_grid_topology

test_triangulation

Triangulations

Interface

Gridap.Geometry.Triangulation — Type

abstract type Triangulation{Dc,Dp}

Abstract type representing an arbitrary tiling, tessellation, or triangulation of a domain of parametric dimension Dc and physical dimension Dp.

We define a triangulation from two basic ingredients:

the cell-wise nodal coordinates of the cells in the triangulation, plus

an interpolation of this cell-wise coordinates into the cells interior.

Note that this type represents general triangulations (not necessarily conforming), which is the minimum geometrical information needed to perform cell-wise numerical integration.

The Triangulation interface is defined by overloading these methods:

get_cell_coordinates(trian::Triangulation)

get_reffes(trian::Triangulation)

get_cell_type(trian::Triangulation)

Optional interface:

For triangulations living in a space of co-dimension 1, the following method can be defined:

[get_normal_vector(trian::Triangulation)]

In some cases, concrete implementations want to override the default implementation of the following methods:

[restrict(f::AbstractArray, trian::Triangulation)]

[get_cell_id(f::AbstractArray, trian::Triangulation)]

The (mandatory) Triangulation interface can be tested with

test_triangulation

Gridap.Geometry.get_cell_coordinates — Method

get_cell_coordinates(trian::Triangulation) -> AbstractArray{Vector{<:Point{Dp}}}

Gridap.Geometry.get_reffes — Method

get_reffes(trian::Triangulation) -> Vector{LagrangianRefFE}

Gridap.Geometry.get_cell_type — Method

get_cell_type(trian::Triangulation) -> AbstractVector{<:Integer}

Gridap.Geometry.get_normal_vector — Method

get_normal_vector(trian::Triangulation)

Gridap.Geometry.restrict — Method

restrict(f::AbstractArray, trian::Triangulation)

Gridap.Arrays.reindex — Method

reindex(a::AbstractArray, trian::Triangulation)

Gridap.Geometry.test_triangulation — Function

test_triangulation(trian::Triangulation)

Gridap.Geometry.num_cells — Method

num_cells(trian::Triangulation) -> Int

Gridap.ReferenceFEs.num_cell_dims — Method

num_cell_dims(::GridTopology) -> Int
num_cell_dims(::Type{<:GridTopology}) -> Int

num_cell_dims(::Triangulation) -> Int
num_cell_dims(::Type{<:Triangulation}) -> Int

Gridap.Integration.num_point_dims — Method

num_point_dims(::GridTopology) -> Int
num_point_dims(::Type{<:GridTopology}) -> Int

num_point_dims(::Triangulation) -> Int
num_point_dims(::Type{<:Triangulation}) -> Int

num_point_dims(::Quadrature{D}) where D
num_point_dims(::Type{<:Quadrature{D}}) where D

Gridap.Integration.num_dims — Method

num_dims(::GridTopology) -> Int
num_dims(::Type{<:GridTopology}) -> Int

Equivalent to num_cell_dims.

num_dims(::Triangulation) -> Int
num_dims(::Type{<:Triangulation}) -> Int

Equivalent to num_cell_dims.

num_dims(::Quadrature{D}) where D where D
num_dims(::Type{<:Quadrature{D}}) where D

Gridap.ReferenceFEs.is_affine — Method

is_affine(trian::Triangulation) -> Bool

Gridap.ReferenceFEs.is_first_order — Method

is_first_order(trian::Triangulation) -> Bool

Gridap.Geometry.get_cell_reffes — Method

get_cell_reffes(trian::Triangulation) -> Vector{<:NodalReferenceFEs}

It is not desirable to iterate over the resulting array for large number of cells if the underlying reference FEs are of different Julia type.

Gridap.Geometry.get_cell_shapefuns — Method

get_cell_shapefuns(trian::Triangulation) -> Vector{<:Field}

Gridap.Geometry.get_cell_map — Method

get_cell_map(trian::Triangulation) -> Vector{<:Field}

Gridap.Geometry.get_physical_coordinate — Method

get_physical_coordinate(trian::Triangulation)

In contrast to getcellmap, the returned object:

is a CellField

its gradient is the identity tensor

Integrating on a triangulation

Gridap.Geometry.CellQuadrature — Type

struct CellQuadrature <: GridapType
  array
end

Gridap.Geometry.CellQuadrature — Method

CellQuadrature(array::AbstractArray{<:Quadrature})

Gridap.Geometry.CellQuadrature — Method

CellQuadrature(trian::Triangulation, degree::Integer)

Gridap.Geometry.CellQuadrature — Method

CellQuadrature(polytopes::Vector{<:Polytope}, cell_types::AbstractVector)

Gridap.Integration.get_coordinates — Method

get_coordinates(q::CellQuadrature)

Gridap.Integration.get_weights — Method

get_weights(q::CellQuadrature)

Gridap.Arrays.get_array — Method

get_array(quad::CellQuadrature)

Gridap.Fields.integrate — Method

integrate(cell_field,trian::Triangulation,quad::CellQuadrature)

The cell_field is aligned with the cells in trian

TriangulationPortion

Gridap.Geometry.TriangulationPortion — Type

Gridap.Geometry.TriangulationPortion — Method

RestrictedTriangulation

Gridap.Geometry.RestrictedTriangulation — Type

BoundaryTriangulations

Interface

Gridap.Geometry.BoundaryTriangulation — Type

abstract type BoundaryTriangulation{Dc,Dp} <: Triangulation{Dc,Dp}

Gridap.Geometry.BoundaryTriangulation — Method

BoundaryTriangulation(model::DiscreteModel,face_to_mask::Vector{Bool})
BoundaryTriangulation(model::DiscreteModel)

Gridap.Geometry.BoundaryTriangulation — Method

BoundaryTriangulation(model::DiscreteModel,tags::Vector{Int})
BoundaryTriangulation(model::DiscreteModel,tags::Vector{String})
BoundaryTriangulation(model::DiscreteModel,tag::Int)
BoundaryTriangulation(model::DiscreteModel,tag::String)

Gridap.Geometry.get_volume_triangulation — Method

get_volume_triangulation(trian::BoundaryTriangulation)

Gridap.Geometry.get_face_to_cell — Method

Gridap.Geometry.get_face_to_lface — Method

Gridap.Geometry.get_face_to_cell_map — Method

Gridap.Geometry.get_normal_vector — Method

Gridap.Geometry.test_boundary_triangulation — Function

test_boundary_triangulation(trian::BoundaryTriangulation)

GenericBoundaryTriangulations

Gridap.Geometry.GenericBoundaryTriangulation — Type

struct GenericBoundaryTriangulation{Dc,Dp,Gf,Gc,O} <: BoundaryTriangulation{Dc,Dp}
  face_trian::Gf
  cell_trian::Gc
  # + private fields
end

Gridap.Geometry.GenericBoundaryTriangulation — Method

GenericBoundaryTriangulation(model::DiscreteModel,face_to_mask::Vector{Bool})

SkeletonTriangulations

Gridap.Geometry.SkeletonTriangulation — Type

struct SkeletonTriangulation{Dc,Dp,B} <: Triangulation{Dc,Dp}
  left::B
  right::B
end

The inner constructor enforces B<:BoundaryTriangulation

Gridap.Geometry.SkeletonTriangulation — Method

SkeletonTriangulation(model::DiscreteModel,face_to_mask::Vector{Bool})
SkeletonTriangulation(model::DiscreteModel)

Gridap.Geometry.InterfaceTriangulation — Method

Gridap.Geometry.get_volume_triangulation — Method

get_volume_triangulation(trian::SkeletonTriangulation)

Gridap.Geometry.get_normal_vector — Method

get_normal_vector(trian::SkeletonTriangulation)

Gridap.Geometry.get_left_boundary — Method

Gridap.Geometry.get_right_boundary — Method

SkeletonPairs

Gridap.Geometry.SkeletonPair — Type

struct SkeletonPair{L,R} <: GridapType
  left::L
  right::R
end

Grids

Interface

Gridap.Geometry.Grid — Type

abstract type Grid{Dc,Dp} <: Triangulation{Dc,Dp}

Abstract type that represents conforming triangulations, whose cell-wise nodal coordinates are defined with a vector of nodal coordinates, plus a cell-wise vector of node ids.

The interface of Grid is defined by overloading the methods in Triangulation plus the following ones:

get_node_coordinates(trian::Grid)

get_cell_nodes(trian::Grid)

From these two methods a default implementation of get_cell_coordinates(trian::Triangulation) is available.

The Grid interface has the following traits

OrientationStyle(::Type{<:Grid})

RegularityStyle(::Type{<:Grid})

The interface of Grid is tested with

test_grid

Gridap.Geometry.OrientationStyle — Method

OrientationStyle(::Type{<:Grid}) -> Val{Bool}
OrientationStyle(::Grid) -> Val{Bool}

Val{true}() if has oriented faces, Val{false}() otherwise (default).

Gridap.Geometry.RegularityStyle — Method

RegularityStyle(::Type{<:Grid}) -> Val{Bool}
RegularityStyle(::Grid) -> Val{Bool}

Val{true}() if no hanging-nodes (refault), Val{false}() otherwise.

Gridap.ReferenceFEs.get_node_coordinates — Method

get_node_coordinates(trian::Grid) -> AbstractArray{<:Point{Dp}}

Gridap.Geometry.get_cell_nodes — Method

get_cell_nodes(trian::Grid)

Gridap.Geometry.test_grid — Function

test_grid(trian::Grid)

Gridap.ReferenceFEs.num_nodes — Method

num_nodes(trian::Grid) -> Int

Gridap.Geometry.is_oriented — Method

is_oriented(::Type{<:Grid}) -> Bool
is_oriented(a::Grid) -> Bool

Gridap.Geometry.is_regular — Method

is_regular(::Type{<:Grid}) -> Bool
is_regular(a::Grid) -> Bool

Gridap.Geometry.Grid — Method

Grid(reffe::LagrangianRefFE)

Gridap.Geometry.compute_linear_grid — Method

compute_linear_grid(reffe::LagrangianRefFE)

Gridap.Geometry.compute_reference_grid — Method

compute_reference_grid(p::LagrangianRefFE, nelems::Integer)

Gridap.Geometry.Grid — Method

Grid(::Type{ReferenceFE{d}},p::Polytope) where d

Gridap.Geometry.GridTopology — Method

GridTopology(grid::Grid)
GridTopology(grid::Grid, cell_to_vertices::Table, vertex_to_node::Vector)

Gridap.ReferenceFEs.simplexify — Method

simplexify(grid::Grid)

UnstructuredGrids

Gridap.Geometry.UnstructuredGrid — Type

struct UnstructuredGrid{Dc,Dp,Tp,Ti,O} <: Grid{Dc,Dp}
  node_coordinates::Vector{Point{Dp,Tp}}
  cell_nodes::Table{Ti,Int32}
  reffes::Vector{<:NodalReferenceFE{Dc}}
  cell_types::Vector{Int8}
end

Gridap.Geometry.UnstructuredGrid — Method

function UnstructuredGrid(
  node_coordinates::Vector{Point{Dp,Tp}},
  cell_nodes::Table{Ti},
  reffes::Vector{<:NodalReferenceFE{Dc}},
  cell_types::Vector,
  ::Val{B}=Val{false}()) where {Dc,Dp,Tp,Ti,B}
end

Low-level inner constructor.

Gridap.Geometry.UnstructuredGrid — Method

UnstructuredGrid(trian::Grid)

Gridap.Geometry.UnstructuredGrid — Method

UnstructuredGrid(reffe::LagrangianRefFE)

Build a grid with a single cell that is the given reference FE itself

Gridap.Geometry.UnstructuredGrid — Method

UnstructuredGrid(::Type{ReferenceFE{d}},p::Polytope) where d

Gridap.Geometry.UnstructuredGrid — Method

UnstructuredGrid(x::AbstractArray{<:Point})

Gridap.Geometry.UnstructuredGridTopology — Method

UnstructuredGridTopology(grid::UnstructuredGrid)

UnstructuredGridTopology(
  grid::UnstructuredGrid,
  cell_to_vertices::Table,
  vertex_to_node::AbstractVector)

CartesianGrids

Gridap.Geometry.CartesianGrid — Type

struct CartesianGrid{D,T,F} <: Grid{D,D}
  # private fields
end

Gridap.Geometry.CartesianGrid — Method

CartesianGrid(desc::CartesianDescriptor)

Gridap.Geometry.CartesianGrid — Method

CartesianGrid(args...)

Same args needed to construct a CartesianDescriptor

Gridap.Geometry.get_cartesian_descriptor — Method

get_cartesian_descriptor(grid::CartesianGrid)

Get the descriptor of the Cartesian grid

Gridap.Geometry.CartesianDescriptor — Type

struct CartesianDescriptor{D,T,F<:Function}
  origin::Point{D,T}
  sizes::Point{D,T}
  partition::Point{D,Int}
  map::F
end

Struct that stores the data defining a Cartesian grid.

Gridap.Geometry.CartesianDescriptor — Method

CartesianDescriptor(origin,sizes,partition,map::Function=identity)

Gridap.Geometry.CartesianDescriptor — Type

CartesianDescriptor(domain,partition,map::Function=identity)

GridPortion

Gridap.Geometry.GridPortion — Type

struct GridPortion{Dc,Dp,G} <: Grid{Dc,Dp}
  oldgrid::G
  cell_to_oldcell::Vector{Int}
  node_to_oldnode::Vector{Int}
end

Gridap.Geometry.GridPortion — Method

GridPortion(oldgrid::Grid{Dc,Dp},cell_to_oldcell::Vector{Int}) where {Dc,Dp}

FaceLabeling

Gridap.Geometry.FaceLabeling — Type

struct FaceLabeling <: GridapType
  d_to_dface_to_entity::Vector{Vector{Int32}}
  tag_to_entities::Vector{Vector{Int32}}
  tag_to_name::Vector{String}
end

Gridap.Geometry.FaceLabeling — Method

FaceLabeling(d_to_num_dfaces::Vector{Int})
FaceLabeling(topo::GridTopology)

Gridap.Integration.num_dims — Method

num_dims(lab::FaceLabeling)

Gridap.ReferenceFEs.num_cell_dims — Method

num_cell_dims(lab::FaceLabeling)

Gridap.Geometry.num_tags — Method

num_tags(lab::FaceLabeling)

Gridap.Geometry.num_entities — Method

num_entities(lab::FaceLabeling)

Gridap.ReferenceFEs.num_faces — Method

num_faces(lab::FaceLabeling,d::Integer)

Gridap.ReferenceFEs.num_faces — Method

num_faces(lab::FaceLabeling)

Gridap.ReferenceFEs.num_vertices — Method

num_vertices(lab::FaceLabeling)

Gridap.ReferenceFEs.num_edges — Method

num_edges(lab::FaceLabeling)

Gridap.ReferenceFEs.num_facets — Method

num_facets(lab::FaceLabeling)

Gridap.Geometry.num_cells — Method

num_cells(lab::FaceLabeling)

Gridap.Geometry.get_face_entity — Method

get_face_entity(lab::FaceLabeling,d::Integer)

Gridap.Geometry.get_face_entity — Method

get_face_entity(lab::FaceLabeling)

Gridap.Geometry.get_cell_entity — Method

Gridap.Geometry.get_tag_entities — Method

get_tag_entities(lab::FaceLabeling,tag::Integer)
get_tag_entities(lab::FaceLabeling,tag::String)

Gridap.Geometry.get_tag_entities — Method

get_tag_entities(lab::FaceLabeling)

Gridap.Geometry.get_tag_name — Method

get_tag_name(lab::FaceLabeling,tag::Integer)

Gridap.Geometry.get_tag_name — Method

get_tag_name(lab::FaceLabeling)

Gridap.Geometry.get_tag_from_name — Method

get_tag_from_name(lab::FaceLabeling,name::String)

Gridap.Geometry.get_tag_from_name — Method

get_tag_from_name(lab::FaceLabeling)

Gridap.Geometry.get_tags_from_names — Method

get_tags_from_names(lab::FaceLabeling,names::Vector{String})

Gridap.Geometry.get_face_mask — Method

get_face_mask(labeling::FaceLabeling,tags::Vector{Int},d::Integer)
get_face_mask(labeling::FaceLabeling,tags::Vector{String},d::Integer)
get_face_mask(labeling::FaceLabeling,tag::Int,d::Integer)
get_face_mask(labeling::FaceLabeling,tag::String,d::Integer)

Gridap.Geometry.add_tag! — Method

add_tag!(lab::FaceLabeling,name::String,entities::Vector{<:Integer})

Gridap.Geometry.add_tag_from_tags! — Method

add_tag_from_tags!(lab::FaceLabeling, name::String, tags::Vector{Int})
add_tag_from_tags!(lab::FaceLabeling, name::String, tags::Vector{String})
add_tag_from_tags!(lab::FaceLabeling, name::String, tag::Int)
add_tag_from_tags!(lab::FaceLabeling, name::String, tag::String)

Gridap.Geometry.get_face_tag — Method

get_face_tag(labeling::FaceLabeling,tags::Vector{Int},d::Integer)
get_face_tag(labeling::FaceLabeling,tags::Vector{String},d::Integer)
get_face_tag(labeling::FaceLabeling,tag::Int,d::Integer)
get_face_tag(labeling::FaceLabeling,tag::String,d::Integer)
get_face_tag(labeling::FaceLabeling,d::Integer)

The first of the given tags appearing in the face is taken. If there is no tag on a face, this face will have a value equal to UNSET. If not tag or tags are provided, all the tags in the model are considered

Gridap.Geometry.get_face_tag_index — Method

get_face_tag_index(labeling::FaceLabeling,tags::Vector{Int},d::Integer)
get_face_tag_index(labeling::FaceLabeling,tags::Vector{String},d::Integer)
get_face_tag_index(labeling::FaceLabeling,tag::Int,d::Integer)
get_face_tag_index(labeling::FaceLabeling,tag::String,d::Integer)

Like get_face_tag by provides the index into the array tags instead of the tag stored in tags.

GridTopology

Interface

Gridap.Geometry.GridTopology — Type

abstract type GridTopology{Dc,Dp}

Abstract type representing the topological information associated with a grid.

The GridTopology interface is defined by overloading the methods:

get_faces(g::GridTopology,dimfrom::Integer,dimto::Integer)

get_polytopes(g::GridTopology)

get_cell_type(g::GridTopology)

get_vertex_coordinates(g::GridTopology)

The GridTopology interface has the following traits

OrientationStyle(::Type{<:GridTopology})

RegularityStyle(::Type{<:GridTopology})

and tested with this function:

test_grid_topology

Gridap.Geometry.OrientationStyle — Method

OrientationStyle(::Type{<:GridTopology}) -> Val{Bool}
OrientationStyle(::GridTopology) -> Val{Bool}

Val{true}() if has oriented faces, Val{false}() otherwise (default).

Gridap.Geometry.RegularityStyle — Method

RegularityStyle(::Type{<:GridTopology}) -> Val{Bool}
RegularityStyle(::GridTopology) -> Val{Bool}

Val{true}() if no hanging-faces (refault), Val{false}() otherwise.

Gridap.ReferenceFEs.get_faces — Method

get_faces(g::GridTopology,dimfrom::Integer,dimto::Integer)

Gridap.Geometry.get_polytopes — Method

get_polytopes(g::GridTopology)

Gridap.Geometry.get_cell_type — Method

get_cell_type(g::GridTopology)

Gridap.ReferenceFEs.get_vertex_coordinates — Method

get_vertex_coordinates(g::GridTopology)

Gridap.Geometry.test_grid_topology — Method

test_grid_topology(top::GridTopology)

Gridap.ReferenceFEs.num_cell_dims — Method

num_cell_dims(::GridTopology) -> Int
num_cell_dims(::Type{<:GridTopology}) -> Int

num_cell_dims(::Triangulation) -> Int
num_cell_dims(::Type{<:Triangulation}) -> Int

Gridap.Integration.num_point_dims — Method

num_point_dims(::GridTopology) -> Int
num_point_dims(::Type{<:GridTopology}) -> Int

num_point_dims(::Triangulation) -> Int
num_point_dims(::Type{<:Triangulation}) -> Int

num_point_dims(::Quadrature{D}) where D
num_point_dims(::Type{<:Quadrature{D}}) where D

Gridap.Integration.num_dims — Method

num_dims(::GridTopology) -> Int
num_dims(::Type{<:GridTopology}) -> Int

Equivalent to num_cell_dims.

num_dims(::Triangulation) -> Int
num_dims(::Type{<:Triangulation}) -> Int

Equivalent to num_cell_dims.

num_dims(::Quadrature{D}) where D where D
num_dims(::Type{<:Quadrature{D}}) where D

Gridap.ReferenceFEs.num_faces — Method

num_faces(g::GridTopology,d::Integer)
num_faces(g::GridTopology)

Gridap.Geometry.num_cells — Method

num_cells(g::GridTopology)

Gridap.ReferenceFEs.num_facets — Method

num_facets(g::GridTopology)

Gridap.ReferenceFEs.num_edges — Method

num_edges(g::GridTopology)

Gridap.ReferenceFEs.num_vertices — Method

num_vertices(g::GridTopology)

Gridap.ReferenceFEs.get_dimranges — Method

get_dimranges(g::GridTopology)

Gridap.ReferenceFEs.get_dimrange — Method

get_dimrange(g::GridTopology,d::Integer)

Gridap.ReferenceFEs.get_offsets — Method

get_offsets(g::GridTopology)

Gridap.ReferenceFEs.get_offset — Method

get_offset(g::GridTopology,d::Integer)

Gridap.ReferenceFEs.get_facedims — Method

get_facedims(g::GridTopology)

Gridap.Geometry.get_cell_faces — Method

get_cell_faces(g::GridTopology)

Defaults to

compute_cell_faces(g)

Gridap.Geometry.compute_cell_faces — Method

compute_cell_faces(g::GridTopology)

Gridap.ReferenceFEs.get_face_vertices — Method

get_face_vertices(g::GridTopology,d::Integer)

Gridap.ReferenceFEs.get_face_vertices — Method

get_face_vertices(g::GridTopology)

Defaults to

compute_face_vertices(g)

Gridap.Geometry.compute_face_vertices — Method

compute_face_vertices(g::GridTopology)

Gridap.Geometry.get_cell_vertices — Method

get_cell_vertices(g::GridTopology)

Gridap.ReferenceFEs.is_simplex — Method

is_simplex(p::GridTopology) -> Bool

Gridap.ReferenceFEs.is_n_cube — Method

is_n_cube(p::GridTopology) -> Bool

Gridap.Geometry.is_oriented — Method

is_oriented(::Type{<:GridTopology}) -> Bool
is_oriented(a::GridTopology) -> Bool

Gridap.Geometry.is_regular — Method

is_regular(::Type{<:GridTopology}) -> Bool
is_regular(a::GridTopology) -> Bool

Gridap.ReferenceFEs.get_reffaces — Method

get_reffaces(::Type{Polytope{d}}, g::GridTopology) where d

By default, it calls to compute_reffaces.

Gridap.ReferenceFEs.get_face_type — Method

get_face_type(g::GridTopology,d::Integer)

By default, it calls to compute_reffaces.

Gridap.Geometry.compute_reffaces — Method

compute_reffaces(::Type{Polytope{d}}, g::GridTopology) where d

Gridap.ReferenceFEs.get_reffaces — Method

get_reffaces(topo::GridTopology)

Gridap.ReferenceFEs.get_face_type — Method

get_face_type(topo::GridTopology)

Gridap.Geometry.get_reffaces_offsets — Method

get_reffaces_offsets(topo::GridTopology)

Gridap.Geometry.compute_reffaces — Method

compute_reffaces(g::GridTopology)

Gridap.Geometry.get_isboundary_face — Method

get_isboundary_face(g::GridTopology)

Gridap.Geometry.get_isboundary_face — Method

get_isboundary_face(g::GridTopology,d::Integer)

Gridap.Geometry.compute_isboundary_face — Method

compute_isboundary_face(g::GridTopology)

Gridap.Geometry.compute_isboundary_face — Method

compute_isboundary_face(g::GridTopology,d::Integer)

Gridap.Geometry.get_cell_permutations — Method

get_cell_permutations(top::GridTopology)

Gridap.Geometry.get_cell_permutations — Method

get_cell_permutations(top::GridTopology,d::Integer)

Gridap.Geometry.compute_cell_permutations — Method

compute_cell_permutations(top::GridTopology)

Gridap.Geometry.compute_cell_permutations — Method

compute_cell_permutations(top::GridTopology,d::Integer)

UnstructuredGridTopology

Gridap.Geometry.UnstructuredGridTopology — Type

struct UnstructuredGridTopology{Dc,Dp,T,O} <: GridTopology{Dc,Dp}
  # private fields
end

Gridap.Geometry.UnstructuredGridTopology — Method

UnstructuredGridTopology(
  vertex_coordinates::Vector{<:Point},
  cell_vertices::Table,
  cell_type::Vector{<:Integer},
  polytopes::Vector{<:Polytope},
  orientation::Val{O}=Val{false}()) where O

Gridap.Geometry.UnstructuredGridTopology — Method

UnstructuredGridTopology(
  vertex_coordinates::Vector{<:Point},
  d_to_dface_vertices::Vector{<:Table},
  cell_type::Vector{<:Integer},
  polytopes::Vector{<:Polytope},
  orientation::Val{O}=Val{false}()) where O

DiscreteModels

Interface

Gridap.Geometry.DiscreteModel — Type

abstract type DiscreteModel{Dc,Dp} <: GridapType

Abstract type holding information about a physical grid, the underlying grid topology, and a labeling of the grid faces. This is the information that typically provides a mesh generator, and it is what one needs to perform a simulation.

The DiscreteModel interface is defined by overloading the methods:

get_grid(model::DiscreteModel)

get_grid_topology(model::DiscreteModel)

get_face_labeling(g::DiscreteModel)

The interface is tested with this function:

test_discrete_model

Gridap.Geometry.get_grid — Method

get_grid(model::DiscreteModel)

Gridap.Geometry.get_grid_topology — Method

get_grid_topology(model::DiscreteModel)

Gridap.Geometry.get_face_labeling — Method

get_face_labeling(g::DiscreteModel)

Gridap.Geometry.test_discrete_model — Method

test_discrete_model(model::DiscreteModel)

Gridap.Integration.num_dims — Method

num_dims(model::DiscreteModel)

Gridap.ReferenceFEs.num_cell_dims — Method

num_cell_dims(model::DiscreteModel)

Gridap.Integration.num_point_dims — Method

num_point_dims(model::DiscreteModel)

Gridap.ReferenceFEs.num_faces — Method

num_faces(g::DiscreteModel,d::Integer)
num_faces(g::DiscreteModel)

Gridap.Geometry.num_cells — Method

num_cells(g::DiscreteModel)

Gridap.ReferenceFEs.num_facets — Method

num_facets(g::DiscreteModel)

Gridap.ReferenceFEs.num_edges — Method

num_edges(g::DiscreteModel)

Gridap.ReferenceFEs.num_vertices — Method

num_vertices(g::DiscreteModel)

Gridap.ReferenceFEs.num_nodes — Method

num_nodes(g::DiscreteModel)

Gridap.ReferenceFEs.get_face_nodes — Method

get_face_nodes(g::DiscreteModel,d::Integer)

Gridap.ReferenceFEs.get_face_nodes — Method

get_face_nodes(g::DiscreteModel)

Gridap.Geometry.compute_face_nodes — Method

compute_face_nodes(model::DiscreteModel,d::Integer)

Gridap.Geometry.compute_face_nodes — Method

compute_face_nodes(model::DiscreteModel)

Gridap.ReferenceFEs.get_face_own_nodes — Method

get_face_own_nodes(g::DiscreteModel,d::Integer)

Gridap.ReferenceFEs.get_face_own_nodes — Method

get_face_own_nodes(g::DiscreteModel)

Gridap.Geometry.compute_face_own_nodes — Method

compute_face_own_nodes(model::DiscreteModel,d::Integer)

Gridap.Geometry.compute_face_own_nodes — Method

compute_face_own_nodes(model::DiscreteModel)

Gridap.ReferenceFEs.get_vertex_node — Method

get_vertex_node(g::DiscreteModel)

Gridap.Geometry.compute_vertex_node — Method

compute_vertex_node(g::DiscreteModel)

Gridap.Geometry.get_node_face_owner — Method

get_node_face_owner(g::DiscreteModel)

Gridap.Geometry.compute_node_face_owner — Method

compute_node_face_owner(g::DiscreteModel)

Gridap.ReferenceFEs.get_reffaces — Method

get_reffaces(::Type{ReferenceFE{d}},model::DiscreteModel) where d

Gridap.ReferenceFEs.get_face_type — Method

get_face_type(g::DiscreteModel,d::Integer)

Index to the vector get_reffaces(ReferenceFE{d},g)

Gridap.Geometry.compute_reffaces — Method

compute_reffaces(::Type{ReferenceFE{d}}, g::DiscreteModel) where d

Gridap.ReferenceFEs.get_reffaces — Method

get_reffaces(model::DiscreteModel)

Gridap.ReferenceFEs.get_face_type — Method

get_face_type(model::DiscreteModel)

Gridap.Geometry.get_reffaces_offsets — Method

get_reffaces_offsets(model::DiscreteModel)

Gridap.Geometry.compute_reffaces — Method

compute_reffaces(g::DiscreteModel)

Gridap.Geometry.Grid — Method

Grid(::Type{ReferenceFE{d}},model::DiscreteModel) where d

Gridap.Geometry.Triangulation — Method

Triangulation(::Type{ReferenceFE{d}},model::DiscreteModel) where d
Triangulation(model::DiscreteModel)

Gridap.Geometry.get_triangulation — Method

get_triangulation(model::DiscreteModel)

Gridap.Geometry.get_polytopes — Method

get_polytopes(model::DiscreteModel)

Gridap.ReferenceFEs.simplexify — Method

simplexify(model::DiscreteModel)

DiscreteModelFromFile

Gridap.Geometry.DiscreteModelFromFile — Method

DiscreteModelFromFile(filename::AbstractString)

UnstructuredDiscreteModels

Gridap.Geometry.UnstructuredDiscreteModel — Type

struct UnstructuredDiscreteModel{Dc,Dp,Tp,B} <: DiscreteModel{Dc,Dp}
  grid::UnstructuredGrid{Dc,Dp,Tp,B}
  grid_topology::UnstructuredGridTopology{Dc,Dp,Tp,B}
  face_labeling::FaceLabeling
end

Gridap.Geometry.UnstructuredDiscreteModel — Method

UnstructuredDiscreteModel(grid::Grid)

CartesianDiscreteModels

Gridap.Geometry.CartesianDiscreteModel — Type

struct CartesianDiscreteModel{D,T,F} <: DiscreteModel{D,D}
  # Private Fields
end

Gridap.Geometry.CartesianDiscreteModel — Method

CartesianDiscreteModel(desc::CartesianDescriptor)

Inner constructor

Gridap.Geometry.CartesianDiscreteModel — Method

CartesianDiscreteModel(args...)

Same args needed to construct a CartesianDescriptor

Gridap.Geometry.get_cartesian_descriptor — Method

get_cartesian_descriptor(model::CartesianDiscreteModel)

DiscreteModelPortion

Gridap.Geometry.DiscreteModelPortion — Type

RestrictedDiscreteModel

Gridap.Geometry.RestrictedDiscreteModel — Type

CellFields

CellFieldLike interface

Gridap.Geometry.CellFieldLike — Type

abstract type CellFieldLike <: GridapType end

Gridap.Arrays.get_array — Method

get_array(cf::CellFieldLike)

Gridap.Geometry.get_cell_map — Method

get_cell_map(cf::CellFieldLike)

Gridap.Geometry.similar_object — Method

similar_object(cf::CellFieldLike,array::AbstractArray)

Gridap.Geometry.similar_object — Method

similar_object(cf1::CellFieldLike,cf2::CellFieldLike,array::AbstractArray)

Gridap.Fields.gradient — Method

gradient(cf::CellFieldLike)

Gridap.Fields.grad2curl — Method

grad2curl(cf::CellFieldLike)

Gridap.Geometry.test_cell_field_like — Function

test_cell_field_like(
  cf::CellFieldLike,
  x::AbstractArray,
  b::AbstractArray,
  pred=(==);
  grad=nothing)

Gridap.Fields.evaluate — Method

evaluate(cf::CellFieldLike,x)

Base.length — Method

length(cf::CellFieldLike)

CellField interface

Gridap.Geometry.CellField — Type

abstract type CellField <: CellFieldLike end

Gridap.Geometry.test_cell_field — Function

test_cell_field(cf::CellField,args...;kwargs...)

Same arguments as test_cell_field_like

Gridap.Geometry.convert_to_cell_field — Method

convert_to_cell_field(object::CellField,cell_map)

Gridap.Geometry.restrict — Method

restrict(cf::CellField,trian::Triangulation)

Concrete implementations

Gridap.Geometry.GenericCellField — Type

struct GenericCellField <: CellField
  array::AbstractArray
  cell_map::AbstractArray
end

Gridap.Geometry.SkeletonCellField — Type

struct SkeletonCellField <: GridapType
  left::CellField
  right::CellField
end

Supports the same differential and algebraic operations than CellField

Gridap.Geometry.QPointCellField — Method

Gridap.Geometry.CellField — Method

Gridap.Geometry.get_cell_map — Method

get_cell_map(a::SkeletonCellField)

Gridap.Geometry.jump — Method

jump(sf::SkeletonCellField)

Gridap.Geometry.mean — Method

mean(sf::SkeletonCellField)