Gridap.MultiField

Gridap.MultiField — Module

The exported names are

ConsecutiveMultiFieldStyle

MultiFieldCellField

MultiFieldFEFunction

MultiFieldFESpace

MultiFieldStyle

compute_field_offsets

interpolate

interpolate_dirichlet

interpolate_everywhere

num_fields

restrict_to_field

Gridap.MultiField.MultiFieldFEFunction — Type

struct MultiFieldFEFunction <: CellField
  # private fields
end

Gridap.MultiField.MultiFieldFESpace — Type

struct MultiFieldFESpace{S<:MultiFieldStyle,B} <: FESpace
  spaces::Vector{<:SingleFieldFESpace}
  multi_field_style::S
  constraint_style::Val{B}
end

Gridap.MultiField.MultiFieldFESpace — Method

MultiFieldFESpace(spaces::Vector{<:SingleFieldFESpace})

Gridap.FESpaces.interpolate — Method

The resulting MultiFieldFEFunction is in the space (in particular it fulfills Dirichlet BCs even in the case that the given cell field does not fulfill them)

Gridap.FESpaces.interpolate_dirichlet — Method

Gridap.FESpaces.interpolate_everywhere — Method

like interpolate, but also compute new degrees of freedom for the dirichlet component. The resulting MultiFieldFEFunction does not necessary belongs to the underlying space

Gridap.MultiField.compute_field_offsets — Method

compute_field_offsets(f::MultiFieldFESpace)

Gridap.MultiField.num_fields — Method

num_fields(m::MultiFieldFEFunction)

Gridap.MultiField.num_fields — Method

num_fields(f::MultiFieldFESpace)

Gridap.MultiField.restrict_to_field — Method

restrict_to_field(f::MultiFieldFESpace,free_values::AbstractVector,field::Integer)